# multitensor shouldn't recompile