--- a/test/test_multitensor.py
+++ b/test/test_multitensor.py
@@ -6,6 +6,7 @@ from tinygrad.ops import LoadOps, ReduceOps
 from tinygrad.helpers import CI, prod, Context
 from tinygrad.nn.state import get_parameters, get_state_dict
 from tinygrad.engine.schedule import create_schedule
+from tinygrad.engine.realize import lower_schedule
 from tinygrad.features.multi import all_reduce, MultiLazyBuffer
 from random import randint
 import numpy as np
@@ -39,6 +40,18 @@ class TestMultiTensor(unittest.TestCase):
     for lb in X.lazydata.lbs:
       assert lb.shape == (128,)
     (X + X).realize()
+
+  def test_shard_no_recompile(self):
+    X = Tensor.ones(256).contiguous().realize()
+    X.shard_((d0, d1), 0)
+    out = (X + X)
+    sched = create_schedule(out.lazydata.lbs)
+    names = []
+    for si, ei in zip(sched[:], lower_schedule(sched)):
+      names.append(ei.prg.name)
+      ei.run()
+    assert names[-4] == names[-3]
+    assert names[-2] == names[-1]
 
   def test_sharded_memory(self):
     # Buffer may be stuck in track_cross_buffer