--- a/tinygrad/device.py
+++ b/tinygrad/device.py
@@ -205,8 +205,7 @@ class MultiDeviceJITGraph(Runner):
   def __call__(self, rawbufs:List[Buffer], var_vals:Dict[Variable, int], wait=False, jit=False) -> Optional[float]:
     raise NotImplementedError("override this")
 
-method_cache: Dict[Tuple[str, Tuple[LazyOp, ...]], CompiledRunner] = {}
-base_method_cache: Dict[Tuple[str, Tuple[LazyOp, ...]], CompiledRunner]  = {}
+method_cache: Dict[Tuple[str, Tuple[LazyOp, ...], bool], CompiledRunner] = {}
 logkern, logkern_level = open(getenv("LOGKERN", ""), "a") if getenv("LOGKERN", "") else None, getenv("LOGKERN_LEVEL", 1)
 class Compiled:
   def __init__(self, device:str, allocator:Allocator, compiler:Optional[Compiler], runtime, graph=None):
@@ -256,10 +255,10 @@ class Compiled:
     return k
 
   def get_runner(self, *ast:LazyOp) -> CompiledRunner:
-    if cret:=method_cache.get((self.dname, ast)): return cret
-    if bret:=base_method_cache.get((self.dname.split(":")[0], ast)):
-      method_cache[(self.dname, ast)] = ret = bret.to_other_device(self.dname)
+    if cret:=method_cache.get((self.dname, ast, False)): return cret
+    if bret:=method_cache.get((self.dname.split(":")[0], ast, True)):
+      method_cache[(self.dname, ast, False)] = ret = bret.to_other_device(self.dname)
     else:
-      base_method_cache[(self.dname.split(":")[0], ast)] = method_cache[(self.dname, ast)] = ret = self.to_program(self.get_linearizer(*ast))
+      method_cache[(self.dname.split(":")[0], ast, True)] = method_cache[(self.dname, ast, False)] = ret = self.to_program(self.get_linearizer(*ast))
     return ret