On master:
```
tiny@tiny14:~/tinygrad$ BS=1536 GPUS=6 python3 examples/hlb_cifar10.py
shuffling training dataset in 2196.98 ms (epoch=0)
  0 24073.57 ms run, 24073.25 ms python,    0.31 ms CUDA * 6, 1182.99 loss, 0.000015 LR, 1.11 GB used,     83.79 GFLOPS,   2017.22 GOPS
  1 8122.71 ms run, 8122.15 ms python,    0.56 ms CUDA * 6, 1181.49 loss, 0.000030 LR, 12.35 GB used,    248.21 GFLOPS,   2016.16 GOPS
  2   62.91 ms run,   23.67 ms python,   39.24 ms CUDA * 6, 1179.67 loss, 0.000045 LR, 12.35 GB used,  32048.85 GFLOPS,   2016.16 GOPS
  3   61.12 ms run,   20.80 ms python,   40.32 ms CUDA * 6, 1171.41 loss, 0.000060 LR, 12.35 GB used,  32984.38 GFLOPS,   2016.16 GOPS
  4   61.12 ms run,   20.59 ms python,   40.52 ms CUDA * 6, 1165.13 loss, 0.000075 LR, 12.35 GB used,  32989.38 GFLOPS,   2016.16 GOPS
```

This branch:
```
tiny@tiny14:~/tinygrad$ BS=1536 GPUS=6 python3 examples/hlb_cifar10.py
shuffling training dataset in 2186.14 ms (epoch=0)
  0 8932.78 ms run, 8932.48 ms python,    0.30 ms CUDA * 6, 1181.03 loss, 0.000015 LR, 1.11 GB used,    225.82 GFLOPS,   2017.22 GOPS
  1 4206.60 ms run, 4206.13 ms python,    0.48 ms CUDA * 6, 1176.76 loss, 0.000030 LR, 12.35 GB used,    479.28 GFLOPS,   2016.16 GOPS
  2   61.18 ms run,   23.17 ms python,   38.01 ms CUDA * 6, 1175.29 loss, 0.000045 LR, 12.35 GB used,  32952.34 GFLOPS,   2016.16 GOPS
  3   60.11 ms run,   18.60 ms python,   41.51 ms CUDA * 6, 1171.62 loss, 0.000060 LR, 12.35 GB used,  33539.83 GFLOPS,   2016.16 GOPS
  4   53.93 ms run,   12.73 ms python,   41.20 ms CUDA * 6, 1168.32 loss, 0.000075 LR, 12.35 GB used,  37381.52 GFLOPS,   2016.16 GOPS
  ```